asta aqui llegue

+++ view/aprendices/delete.php
@@ -1,5 +1,10 @@
 <?php
-    require_once("C://xampp/htdocs/CRUD_APRENDICES/controller/aprendizController.php");
+require_once("C://xampp/htdocs/CRUD_APRENDICES/controller/aprendizController.php");
+
+if (isset($_GET['id'])) {
     $aprendiz = new AprendizController();
     $aprendiz->delete($_GET['id']);
+} else {
+    echo "ID no proporcionado";
+}
 ?>
